backend/requirements.txt
Django==2.2.3
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
ptvsd==4.2.10
